+++ packages/snap-preact-components/src/components/Organisms/Results/Results.tsx
@@ -18,10 +18,34 @@ import { useDisplaySettings } from '../../../hooks/useDisplaySettings';
 const CSS = {
 	results: ({ columns, gapSize }) =>
 		css({
-			display: 'grid',
-			gridTemplateColumns: `repeat(${columns}, 1fr)`,
-			gridTemplateRows: 'auto',
+			display: 'flex',
+			flexFlow: 'row wrap',
 			gap: gapSize,
+			gridTemplateRows: 'auto',
+			gridTemplateColumns: `repeat(${columns}, 1fr)`,
+
+			'& .ss__result': {
+				boxSizing: 'border-box',
+				flex: '0 1 auto',
+				width: `calc(${100 / columns}% - (${columns - 1} * ${gapSize} / ${columns} ) )`,
+				marginRight: gapSize,
+				marginBottom: gapSize,
+
+				[`&:nth-of-type(${columns}n)`]: {
+					marginRight: '0',
+				},
+				[`&:nth-last-of-type(-n+${columns})`]: {
+					marginBottom: '0',
+				},
+			},
+			'@supports (display: grid)': {
+				display: 'grid',
+
+				'& .ss__result': {
+					width: 'initial',
+					margin: 0,
+				},
+			},
 		}),
 };