+++ packages/snap-preact/src/Instantiators/RecommendationInstantiator.tsx
@@ -16,6 +16,8 @@ export type RecommendationInstantiatorConfig = {
 	};
 	config: {
 		branch: string;
+		realtime?: boolean;
+		batched?: boolean;
 	} & Attachments;
 	selector?: string;
 	services?: SnapControllerServices;
@@ -86,24 +88,43 @@ export class RecommendationInstantiator {
 			async (target, injectedElem, elem) => {
 				const globals: any = {};
 
-				const { shopper, shopperId, product, seed, branch, options } = getContext(
-					['shopperId', 'shopper', 'product', 'seed', 'branch', 'options'],
+				const { shopper, shopperId, product, seed, options } = getContext(
+					['shopperId', 'shopper', 'product', 'seed', 'options'],
 					elem as HTMLScriptElement
 				);
 
+				/*
+					type instantiatorContext = {
+						shopper?: {
+							id?: string;
+							cart?: Array<{ sku: string; }>;
+						};
+						shopperId?: string;
+						product?: string;
+						seed?: string;
+						options?: {
+							siteId?: string;
+							branch?: string;
+							batched?: boolean;
+							realtime?: boolean;
+							categories?: any;
+						}
+					}
+				*/
+
 				if (shopper || shopperId) {
-					globals.shopper = shopper || shopperId;
+					globals.shopper = shopper?.id || shopperId;
 				}
 				if (product || seed) {
 					globals.product = product || seed;
 				}
-				if (branch) {
-					globals.branch = branch;
+				if (options?.branch) {
+					globals.branch = options.branch;
 				}
-				if (options && options.siteId) {
+				if (options?.siteId) {
 					globals.siteId = options.siteId;
 				}
-				if (options && options.categories) {
+				if (options?.categories) {
 					globals.categories = options.categories;
 				}
 
@@ -113,9 +134,12 @@ export class RecommendationInstantiator {
 				const controllerConfig = {
 					id: `recommend_${tag + (profileCount[tag] - 1)}`,
 					tag,
+					batched: options?.batched ?? true,
+					realtime: Boolean(options?.realtime),
 					globals,
 					...this.config.config,
 				};
+
 				const createRecommendationController = (await import('../create/createRecommendationController')).default;
 				const client = this.config.services?.client || this.client;
 				const tracker = this.config.services?.tracker || this.tracker;