+++ packages/snap-store-mobx/src/types.ts
@@ -77,6 +77,8 @@ export type RecommendationStoreConfig = StoreConfig & {
 	globals?: any;
 	tag: string;
 	branch?: string;
+	realtime?: boolean;
+	batched?: boolean;
 };
 
 export type StoreConfigs = SearchStoreConfig | AutocompleteStoreConfig | FinderStoreConfig | RecommendationStoreConfig;