+++ packages/snap-toolbox/src/index.ts
@@ -5,3 +5,4 @@ export * from './cookies/cookies';
 export * from './getContext/getContext';
 export * from './url/url';
 export * from './version/version';
+export * from './charsParams/charsParams';