+++ packages/snap-tracker/src/Tracker.ts
@@ -2,7 +2,7 @@ import deepmerge from 'deepmerge';
 import { v4 as uuidv4 } from 'uuid';
 
 import { StorageStore, StorageType } from '@searchspring/snap-store-mobx';
-import { cookies, featureFlags, version, DomTargeter, getContext } from '@searchspring/snap-toolbox';
+import { cookies, featureFlags, version, DomTargeter, getContext, charsParams } from '@searchspring/snap-toolbox';
 
 import { TrackEvent } from './TrackEvent';
 import { PixelEvent } from './PixelEvent';
@@ -20,6 +20,8 @@ import {
 	ShopperLoginEvent,
 	OrderTransactionData,
 	Product,
+	TrackerConfig,
+	PreflightRequestModel,
 } from './types';
 
 const BATCH_TIMEOUT = 150;
@@ -30,39 +32,51 @@ const VIEWED_COOKIE_EXPIRATION = 220752000000; // 7 years
 const COOKIE_SAMESITE = 'Lax';
 const SESSIONID_STORAGE_NAME = 'ssSessionIdNamespace';
 const LOCALSTORAGE_BEACON_POOL_NAME = 'ssBeaconPool';
+const CART_PRODUCTS = 'ssCartProducts';
 const VIEWED_PRODUCTS = 'ssViewedProducts';
 const MAX_VIEWED_COUNT = 15;
-const CART_PRODUCTS = 'ssCartProducts';
+
+const defaultConfig: TrackerConfig = {
+	id: 'track',
+};
 
 export class Tracker {
 	globals: TrackerGlobals;
 	localStorage: StorageStore;
-	sessionStorage: StorageStore;
 	context: BeaconContext;
 	isSending: number;
-	namespace = '';
+
+	private config: TrackerConfig;
 	private targeters: DomTargeter[] = [];
 
-	constructor(globals: TrackerGlobals) {
+	constructor(globals: TrackerGlobals, config?: TrackerConfig) {
 		if (typeof globals != 'object' || typeof globals.siteId != 'string') {
 			throw new Error(`Invalid config passed to tracker. The "siteId" attribute must be provided.`);
 		}
 
+		this.config = deepmerge(defaultConfig, config || {});
+
 		this.globals = globals;
-		this.setNamespace();
+
+		this.localStorage = new StorageStore({
+			type: StorageType.LOCAL,
+			key: `ss-${this.config.id}-${this.globals.siteId}-local`,
+		});
 
 		this.context = {
-			...this.getUserId(),
-			...this.getSessionId(),
-			...this.getShopperId(),
+			userId: this.getUserId(),
+			sessionId: this.getSessionId(),
+			shopperId: this.getShopperId(),
 			pageLoadId: uuidv4(),
 			website: {
 				trackingCode: this.globals.siteId,
 			},
 		};
 
-		if (!window.searchspring?.track) {
-			this.setGlobal();
+		if (!window.searchspring?.tracker) {
+			window.searchspring = window.searchspring || {};
+			window.searchspring.tracker = this;
+			window.searchspring.version = version;
 		}
 
 		// one targeter to rule them all
@@ -90,6 +104,50 @@ export class Tracker {
 			})
 		);
 
+		document.addEventListener('click', (event: Event) => {
+			const attributes = {};
+			Object.values((event.target as HTMLElement).attributes).forEach((attr: Attr) => {
+				attributes[attr.nodeName] = (event.target as HTMLElement).getAttribute(attr.nodeName);
+			});
+
+			const updateRecsControllers = (): void => {
+				if (window.searchspring.controller) {
+					Object.keys(window.searchspring.controller).forEach((name) => {
+						const controller = window.searchspring.controller[name];
+						if (controller.type === 'recommendation' && controller.config?.realtime) {
+							controller.search();
+						}
+					});
+				}
+			};
+
+			if (attributes[`ss-${this.config.id}-cart-add`]) {
+				// add skus to cart
+				const skus = attributes[`ss-${this.config.id}-cart-add`].split(',');
+				this.cookies.cart.add(skus);
+				updateRecsControllers();
+			} else if (attributes[`ss-${this.config.id}-cart-remove`]) {
+				// remove skus from cart
+				const skus = attributes[`ss-${this.config.id}-cart-remove`].split(',');
+				this.cookies.cart.remove(skus);
+				updateRecsControllers();
+			} else if (`ss-${this.config.id}-cart-clear` in attributes) {
+				// clear all from cart
+				this.cookies.cart.clear();
+				updateRecsControllers();
+			} else if (attributes[`ss-${this.config.id}-intellisuggest`] && attributes[`ss-${this.config.id}-intellisuggest-signature`]) {
+				// product click
+				const intellisuggestData = attributes[`ss-${this.config.id}-intellisuggest`];
+				const intellisuggestSignature = attributes[`ss-${this.config.id}-intellisuggest-signature`];
+				const href = attributes['href'];
+				this.track.product.click({
+					intellisuggestData,
+					intellisuggestSignature,
+					href,
+				});
+			}
+		});
+
 		this.sendEvents();
 	}
 
@@ -98,28 +156,6 @@ export class Tracker {
 			target.retarget();
 		});
 	}
-
-	setNamespace = (namespace?: string): void => {
-		let prefix = 'tracker';
-		if (namespace) {
-			this.namespace = `${namespace}`;
-			prefix = namespace;
-		}
-		this.localStorage = new StorageStore({
-			type: StorageType.LOCAL,
-			key: `ss-${prefix}-${this.globals.siteId}-local`,
-		});
-		this.sessionStorage = new StorageStore({
-			type: StorageType.SESSION,
-			key: `ss-${prefix}-${this.globals.siteId}-session`,
-		});
-	};
-
-	setGlobal = (): void => {
-		window.searchspring = window.searchspring || {};
-		window.searchspring.track = this.track;
-		window.searchspring.version = version;
-	};
 
 	track: TrackMethods = {
 		event: (payload: BeaconPayload): BeaconEvent => {
@@ -157,13 +193,13 @@ export class Tracker {
 						},
 					});
 				}
-				const storedShopperId = this.getShopperId()?.shopperId;
+				const storedShopperId = this.getShopperId();
 				data.id = `${data.id}`;
 				if (storedShopperId != data.id) {
 					// user's logged in id has changed, update shopperId cookie send login event
 					cookies.set(SHOPPERID_COOKIE_NAME, data.id, COOKIE_SAMESITE, COOKIE_EXPIRATION);
 					this.context.shopperId = data.id;
-
+					this.sendPreflight();
 					const payload = {
 						type: BeaconType.LOGIN,
 						category: BeaconCategory.PERSONALIZATION,
@@ -203,11 +239,14 @@ export class Tracker {
 				};
 
 				// save recently viewed products to cookie
-				if (data?.sku || data?.childSku) {
-					const viewedProducts = cookies.get(VIEWED_PRODUCTS);
-					const products = viewedProducts ? new Set(viewedProducts.split(',')) : new Set();
-					products.add(data?.sku || data.childSku);
-					cookies.set(VIEWED_PRODUCTS, Array.from(products).slice(0, MAX_VIEWED_COUNT).join(','), COOKIE_SAMESITE, VIEWED_COOKIE_EXPIRATION);
+				const sku = data?.sku || data?.childSku;
+				if (sku) {
+					const lastViewedProducts = this.cookies.viewed.get();
+					const uniqueCartItems = Array.from(new Set([...lastViewedProducts, sku])).map((item) => item.trim());
+					cookies.set(VIEWED_PRODUCTS, uniqueCartItems.slice(0, MAX_VIEWED_COUNT).join(','), COOKIE_SAMESITE, VIEWED_COOKIE_EXPIRATION);
+					if (!lastViewedProducts.includes(sku)) {
+						this.sendPreflight();
+					}
 				}
 
 				// legacy tracking
@@ -303,9 +342,8 @@ export class Tracker {
 
 				// save cart items to cookie
 				if (items.length) {
-					const products = [];
-					items.map((item) => products.push(item.sku || item.childSku));
-					cookies.set(CART_PRODUCTS, products.join(','), COOKIE_SAMESITE, 0);
+					const products = items.map((item) => item.sku || item.childSku);
+					this.cookies.cart.add(products);
 				}
 
 				// legacy tracking
@@ -366,6 +404,9 @@ export class Tracker {
 					event: eventPayload,
 				};
 
+				// clear cart items from cookie when order is placed
+				this.cookies.cart.clear();
+
 				// legacy tracking
 				new PixelEvent(payload);
 
@@ -374,30 +415,31 @@ export class Tracker {
 		},
 	};
 
-	getUserId = (): Record<string, string> => {
+	getUserId = (): string => {
 		let userId;
 		try {
-			userId = featureFlags.storage && this.localStorage.get(USERID_COOKIE_NAME);
+			userId = featureFlags.storage && window.localStorage.getItem(USERID_COOKIE_NAME);
 			if (featureFlags.cookies) {
 				userId = userId || cookies.get(USERID_COOKIE_NAME) || uuidv4();
 				cookies.set(USERID_COOKIE_NAME, userId, COOKIE_SAMESITE, COOKIE_EXPIRATION);
 			} else if (!userId && featureFlags.storage) {
 				// if cookies are disabled, use localStorage instead
 				userId = uuidv4();
-				this.localStorage.set(USERID_COOKIE_NAME, userId);
+				window.localStorage.setItem(USERID_COOKIE_NAME, userId);
 			}
 		} catch (e) {
 			console.error('Failed to persist user id to cookie or local storage:', e);
 		}
-		return { userId };
+
+		return userId;
 	};
 
-	getSessionId = (): Record<string, string> => {
+	getSessionId = (): string => {
 		let sessionId;
 		if (featureFlags.storage) {
 			try {
-				sessionId = this.sessionStorage.get(SESSIONID_STORAGE_NAME) || uuidv4();
-				this.sessionStorage.set(SESSIONID_STORAGE_NAME, sessionId);
+				sessionId = window.sessionStorage.getItem(SESSIONID_STORAGE_NAME) || uuidv4();
+				window.sessionStorage.setItem(SESSIONID_STORAGE_NAME, sessionId);
 				featureFlags.cookies && cookies.set(SESSIONID_STORAGE_NAME, sessionId, COOKIE_SAMESITE, 0); //session cookie
 			} catch (e) {
 				console.error('Failed to persist session id to session storage:', e);
@@ -410,31 +452,123 @@ export class Tracker {
 				cookies.set(SESSIONID_STORAGE_NAME, sessionId, COOKIE_SAMESITE, 0);
 			}
 		}
-		return { sessionId };
+
+		return sessionId;
 	};
 
-	getShopperId = (): Record<string, string> => {
+	getShopperId = (): string => {
 		const shopperId = cookies.get(SHOPPERID_COOKIE_NAME);
 		if (!shopperId) {
 			return;
 		}
-		return { shopperId };
+
+		return shopperId;
 	};
 
-	getCartItems = (): string[] => {
-		const items = cookies.get(CART_PRODUCTS);
-		if (!items) {
-			return [];
+	sendPreflight = (): void => {
+		const userId = this.getUserId();
+		const siteId = this.context.website.trackingCode;
+		const shopper = this.getShopperId();
+		const cart = this.cookies.cart.get();
+		const lastViewed = this.cookies.viewed.get();
+
+		if (userId && siteId && (shopper || cart.length || lastViewed.length)) {
+			const preflightParams: PreflightRequestModel = {
+				userId,
+				siteId,
+			};
+
+			let queryStringParams = `?userId=${encodeURIComponent(userId)}&siteId=${encodeURIComponent(siteId)}`;
+			if (shopper) {
+				preflightParams.shopper = shopper;
+				queryStringParams += `&shopper=${encodeURIComponent(shopper)}`;
+			}
+			if (cart.length) {
+				preflightParams.cart = cart;
+				queryStringParams += cart.map((item) => `&cart=${encodeURIComponent(item)}`).join('');
+			}
+			if (lastViewed.length) {
+				preflightParams.lastViewed = lastViewed;
+				queryStringParams += lastViewed.map((item) => `&lastViewed=${encodeURIComponent(item)}`).join('');
+			}
+
+			const endpoint = `https://${siteId}.a.searchspring.io/api/personalization/preflightCache`;
+			const xhr = new XMLHttpRequest();
+
+			if (charsParams(preflightParams) > 1024) {
+				xhr.open('POST', endpoint);
+				xhr.setRequestHeader('Content-Type', 'application/json');
+				xhr.send(JSON.stringify(preflightParams));
+			} else {
+				xhr.open('GET', endpoint + queryStringParams);
+				xhr.send();
+			}
 		}
-		return items.split(',');
 	};
 
-	getLastViewedItems = (): string[] => {
-		const items = cookies.get(VIEWED_PRODUCTS);
-		if (!items) {
-			return [];
-		}
-		return items.split(',');
+	cookies = {
+		cart: {
+			get: (): string[] => {
+				const items = cookies.get(CART_PRODUCTS);
+				if (!items) {
+					return [];
+				}
+				return items.split(',');
+			},
+			set: (items: string[]): void => {
+				if (items.length) {
+					const cartItems = items.map((item) => item.trim());
+					const uniqueCartItems = Array.from(new Set(cartItems));
+					cookies.set(CART_PRODUCTS, uniqueCartItems.join(','), COOKIE_SAMESITE, 0);
+
+					const itemsHaveChanged = cartItems.filter((item) => items.includes(item)).length !== items.length;
+					if (itemsHaveChanged) {
+						this.sendPreflight();
+					}
+				}
+			},
+			add: (items: string[]): void => {
+				if (items.length) {
+					const currentCartItems = this.cookies.cart.get();
+					const itemsToAdd = items.map((item) => item.trim());
+					const uniqueCartItems = Array.from(new Set([...currentCartItems, ...itemsToAdd]));
+					cookies.set(CART_PRODUCTS, uniqueCartItems.join(','), COOKIE_SAMESITE, 0);
+
+					const itemsHaveChanged = currentCartItems.filter((item) => itemsToAdd.includes(item)).length !== itemsToAdd.length;
+					if (itemsHaveChanged) {
+						this.sendPreflight();
+					}
+				}
+			},
+			remove: (items: string[]): void => {
+				if (items.length) {
+					const currentCartItems = this.cookies.cart.get();
+					const itemsToRemove = items.map((item) => item.trim());
+					const updatedItems = currentCartItems.filter((item) => !itemsToRemove.includes(item));
+					cookies.set(CART_PRODUCTS, updatedItems.join(','), COOKIE_SAMESITE, 0);
+
+					const itemsHaveChanged = currentCartItems.length !== updatedItems.length;
+					if (itemsHaveChanged) {
+						this.sendPreflight();
+					}
+				}
+			},
+			clear: () => {
+				if (this.cookies.cart.get().length) {
+					cookies.unset(CART_PRODUCTS);
+					this.sendPreflight();
+				}
+			},
+		},
+		viewed: {
+			get: (): string[] => {
+				const items = cookies.get(VIEWED_PRODUCTS);
+				if (!items) {
+					return [];
+				}
+				return items.split(',');
+			},
+		},
 	};
 
 	sendEvents = (eventsToSend?: BeaconEvent[]): void => {