+++ packages/snap-tracker/src/types.ts
@@ -4,6 +4,10 @@ export type TrackerGlobals = {
 	siteId: string;
 };
 
+export type TrackerConfig = {
+	id: string;
+};
+
 export interface BeaconPayload {
 	type: BeaconType;
 	category: BeaconCategory;
@@ -138,6 +142,14 @@ export interface ProductClickEvent {
 export interface CustomBeaconEvent {
 	[key: string]: any;
 }
+
+export type PreflightRequestModel = {
+	userId: string;
+	siteId: string;
+	shopper?: string;
+	cart?: string[];
+	lastViewed?: string[];
+};
 
 export interface TrackMethods {
 	event: (payload: BeaconPayload) => BeaconEvent;